Rename py script
